refactor: update getIFrameUrl to use standard web APIs for URLs
